ptxd_dgen_map_all: don't depend on max command line length
Author: Rabin Vincent <rabin@rab.in>

If the command line is larger than the system's limit, xargs will break
the command into multiple executions.  If this occurs in the call to
ptxd_lib_dgen.awk in ptxd_dgen_map_all(), ptxd_lib_dgen.awk will
generate incomplete files because the script's state is not preserved
across the multiple invocations.

This patch avoids this by xargs-ing the filenames to cat before piping
the output to ptxd_lib_dgen.awk.

This was noticed on RHEL4 (which runs a 2.6.9 kernel).  This does not
occur on newer kernels because the limitation on the maximum length of a
command line was removed in 2.6.23.

Signed-off-by: Rabin Vincent <rabin@rab.in>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>

git-svn-id: https://svn.pengutronix.de/svn/ptxdist/trunks/ptxdist-trunk@10186 33e552b5-05e3-0310-8538-816dae2090ed