Provide a choice to select squashfs{3.4,4.0}
linux-2.6.29 now includes squashfs 4.0 in the main line.
It is incompatible with the previous version squashfs3.4

Select 4.0 if you are using linux-2.6.29 or later
Otherwise select the version that matches the squashfs patch you've
applied to your linux kernel tree.

You no longer need to specify the endianess on squashfs4.0

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>

git-svn-id: https://svn.pengutronix.de/svn/ptxdist/trunks/ptxdist-trunk@10200 33e552b5-05e3-0310-8538-816dae2090ed