[canutils] add option to select between toolchain's and BSP's kernel headers
The canutils need the linux/can.h header from the kernel. Normally this
header is provided by the toolchain. If the user is using a toolchain
based on a kernel < 2.6.25 the header is missing.

In the past PTXdist uses the BSP's kernel headers instead, but with
kernel > 2.6.30 this may breaks with the following error message:

    "error: asm/bitsperlong.h: No such file or directory"

So we give the user (or rather the BSP integrator) the choice to select
between these two.....

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>