[libmad] fix patching file order
Makefile.in was patched prior to aclocal.m4. If the gap between the
access timestamps is big enough to let make think aclocal.m4 is newer
than Makefile.in, it will try to regenerate Makefile.in by calling
automake during compiling, which will most certainly fail. So we
switch the file order in the patch to fix this problem.

Signed-off-by: Luotao Fu <l.fu@pengutronix.de>