[lm_sensors] pass MACHINE to makefile in order to overwrite uname -m
This patch fixed the problem that the Makefile uses $(uname -m) to
detect the "host" machine (in autoconf terms). Several x86 binaries are
tried to build if the "build" machine (in autoconf terms) is x86 or
amd64, which will fail if the "host" is e.g. powerpc and lacking the
"sys/io.h" header.

Also don't install "isaset" "isadump" on non x86 platforms.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>