[kernel] install kernel headers into KERNEL_HEADERS_DIR
This patch improves the install stage of the kernel. It installs the
headers with "headers_install" (if available) to KERNEL_HEADERS_DIR, it
will fall back to the old plain copy method.

The headers will be found in $(KERNEL_HEADERS_DIR)/include or simply
$(KERNEL_HEADERS_INCLUDE_DIR). Other packages can use these variables
without having to care about using '=' in assignments.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>