[host-qt4] change bool -> tristate
If a package is bool and it is selected by a "m" module package it turns
to "y". This means when using collections even if qt4 not active ("m"
and not selected by the collection) host-qt4 is turned to "y" and thus
compiled.

So this patch turns the bool into tristate.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>