[bootloader] better "special" compiler handling for bootloader
This patch makes the special compiler handling for bootloaders more
generic. To understand what it does, we look at the current status:

There is the config symbol "COMPILER_PREFIX_UBOOT", if its value
differs from the generic "COMPILER_PREFIX"'s value a special compiler
for u-boot is requested.

Currently only u-boot-v2 uses this feature. So if "U_BOOT_V2" is
active we look if ".utoolchain/$(PTXCONF_COMPILER_PREFIX_UBOOT)gcc"
can be found. If not we ask the user to create the link.

The reader might notice that all other configuration links are named
"selected_*" and that this feature is only limited to u-boot-v2.

To make this more generic, this patch first renames "COMPILER_PREFIX_UBOOT"
into "COMPILER_PREFIX_BOOTLOADER", then a new symbol "BOOTLOADER" is
introduced. All bootloaders that support the new sheme should select
"BOOTLOADER". Now a selected "BOOTLOADER" activates the check for the
special toolchain. The link is renamed to
"selected_toolchain_bootloader" to fit the current scheme.

The bootloaders can use the variable "BOOTLOADER_CROSS_COMPILE" which
can be directly passed to kbuild's CROSS_COMPILE.

A migration handler ist added, so the required symbol conversion is done
automatically by "ptxdist migrate".

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>