[libblkid] Added virtual package
Currently E2FSPROGS is selected for those target packages, which need
libuuid.so or libblkid.so.

The util-linux-ng package contains newer versions of these libraries.

So I added a virtual package 'libblkid', which selects the
appropriate source package.

Later this can easily changed to a real package or redirected to
some other package.

Signed-off-by: Carsten Schlote <c.schlote@konzeptpark.de>
Acked-by: Marc Kleine-Budde <mkl@penguttronix.de>