[host-python] don't interfere with distribution's python
In order to cross compile python, ptxdist will build a special version
of python (i.e. host-python). It is installed to SYSROOT_HOST and the
bin directory is in the PATH.

The host-ipkg-utils need python, too. But we don't want to build
host-python for that, we want to use the distribution's python.

So we delete the executable "python" that is installed by the
"host-python" package. The cross compiler python wrapper will only use
the "python-$VERSION" executable.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>