[ipkg] Updated ipkg.conf config options
Updated default source line for ipkg.conf. It
uses the same sub-directory structure as
ipkg-push would create in the ipkg-repository
for this project.

Added new config option to configure an
ipkg update host. This config can be changed
easily and is used as part of the new default
ipkg.conf entry.

Signed-off-by: Carsten Schlote <c.schlote@konzeptpark.de>
Acked-by: Marc Kleine-Budde <mkl@pengutronix.de>