[image_cpio] generate "images/root.cpio" and "images/root.cpio.gz"
This patch changes the name of the cpio image from "images/initrd.gz" to
"images/root.cpio.gz" which describes the contents better and matches
nameing sheme of the other images.

Further an uncompressed image "image/root.cpio" is introduced.

The meaning of the config symbols changed, too, matching now the file
which are generated:
- IMAGE_CPIO    generates: images/root.cpio
- IMAGE_CPIO_GZ generates: images/root.cpio.gz

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>