[kernel] prepare for "late" installation of kernel image
This patch moves the installation of the kernel image from the
targetinstall stage into the images stage.

This is a preparation for the "whole rootfs in initrd in kernel image"
use case, where first the root file system is build and bundled into a
cpio and then linked into the kernel image.

We define some new symbols:
- IMAGE_KERNEL:
  It's automatically activated if the kernel is build. When active the
  kernel is install to "images/linuximage".
- IMAGE_KERNEL_INSTALL_EARLY:
  default "y". When active the kernel is copied to "images/linuximage"
  during targetinstall.
- IMAGE_KERNEL_INSTALL_LATE:
  If selected the kernel is installed later, during the "images" stage.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>