[kernel/image_kernel] add "whole rootfs in initrd in kernel image" feature
This patch add the "whole rootfs in initrd in kernel image" feature to
the kernel. The whole root file system in form of an cpio image should
be linked into the kernel.

The procedure is split into two parts:
If the feature is active, during the kernel's compile stage an empty
file is linked into the kernel. Because:
a) the kernel should be compiled in the compile stage
b) the root file system isn't read yet, because it is usually not
   compiled. Maybe we even want to put modules into the rootfs.

The second part happens during the "images" stage. First the cpio is
created, then it is linked into the kernel image. Last but not least the
kernel image is installed to "images/kernel" as usual.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>