[host-xorg-server] better fix to avoid patching
This is a improved version of  79c90508b9749b5bcd7d33ed461c6dad46615d40.

By defining "HOST_XORG_SERVER = host-$(XORG_SERVER)" patching and
autogen_gen will look for patches in "..../patches/host-$(XORG_SERVER)" as
we don't have that directory we're not patched and the autogen_gen
script woudn'd create a dependency.

That dependency would cause a circular dep loop, because host-xorg-server
provides m4 files and this is a dependee for "$(STATEDIR)/autogen-tools".

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>