[ptxd_make_import] imake import to different location possible
This patch adds the possibility im import kconfig tree to a different
location, than to the default ${PTXDIST_TOPDIR}/config.

The oackage must setup the variable <PACKAGE>_KCONFIG
that points to the toplevel KCONFIG file.

The imported kconfig tree will be found in
$(PTXDIST_IMPORTDIR)/config/<package>
All symbols are prefixed with <PACKAGE>.

If PTXDIST_IMPORTDIR is omitted, PTXDIST_TOPLEVEL will be used.

the import is started with:
$ ptxdist make <package>_import [PTXDIST_IMPORTDIR=<dir>]

Tested-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>