[ptxd_make_xpkg_pkg] fix installation of read only files
THis patch fixes a problem if a shared lib or an executable is read only
and should be stripped. Some combinations of "install", "fakeroot" and
"glibc" cause the file to be installed read only outside of fakeroot,
others magically change the permissions outside of fakeroot to rw.

However, this means we have first to install the file with rw
permissions, strip it and then change the permissions to read only ones.

Reported-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>