[install_alternative] optional DST for install_alternative
This patch add a new option to the install_alternative call.

Currently install_alternative officially takes 5 parameter:

But many calls use a 6th. parameter: "strip".
This patch first adds a support for the 6th parameter, but marks it as
obsolete. Then a 7th parameter is added, which is an optional destination
for the file.

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>