[sstrip] add option to strip rootfs with sstrip
This patch adds the option to strip the executables and libs of the rot
file system with sstrip instead of binutils strip.

It should be noted that the resulting file cannot be used with gdb,
objdump, or any other program based upon the bfd library, at all. In
fact, the program will not even recognize the file as a valid executable.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>