[ptxdist] rework version generation
This patch aims to fix some problems of the version handling of
ptxdist which were introducted with switching to timed releases:

a) The version is specified in "configure.ac", the "local version"
   ("00003-gbf29f16") is generated by "setlocalversion".
   "setlocalversion" generates it's ouput relative to the latest tag.
   This leads to strange version strings, after creating a tag, and
   forgetting to run "./autogen.sh && ./configure && make".
b) The ptxdist created with a "make dist", even if a not tagged version,
   always thinks it's a release.
c) BSPs working with "master" need a migrate after each release. Because
   the compatibility checker uses year and month.

The problems are^Wshould be fixed with these changes:

a) "setlocalversion" now generates the whole full version, e.g.:
   "2010.04.0-00003-gbf29f16".
   "configure.ac" uses "m4_esyscmd" to figure out the current version
   while running "configure". (Yes, "configure").
b) A file ".tarball-version" is added to the tarball to track the
   version while creating the tarball (same goes for the installation).
   "setlocalversion" will use the version information from
   ".tarball-version"
c) If both ptxdist and config file are not using a release, i.e. their
   versions an "appendix", we say it's compatibe, too.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>