[ptxd_make_world_patchin] rework generation of git "patches" compat dir
For various reasons (git quiltimport cannot handle compressed patches
and uses a hardcoded "series" file) ptxdist creates a special "patches"
folder that git quiltimport" works with.

This patch changes the name of the folder from ".ptxdist/patches" to
".ptxdist/git-patches" so that the original patch folder can always be
found in ".ptxdist/patches" and "patches" (which links to it) regardless
of using git or not.

This fixes a problem with the package "directfb-examples".

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>