[timezone] add new packet
Add an alternative source for the timezone info. This option can be used when
using toolchains in which the timezone info is incorrect/does not work or is
not available at all

The option build's the timezone database from the tzdata source files on the
host and installs the selected zoneinfos to the target.

Signed-off-by: Bart vdr. Meulen <bartvdrmeulen@gmail.com>
Signed-off-by: Remy Bohmer <linux@bohmer.net>