[ptxdist] add data partition to UBI Image
This one adds the option to create a data partition. Currently it is created as
an empty partition. It's considerable that we can change this so we can generate
it with some predefined content. The partition can either be create as a futher
volume in the main root.ubi file or as a standalone flashable UBI image.

Signed-off-by: Luotao Fu <l.fu@pengutronix.de>