Fix dependencies to files in projectroot
When files in the projectroot directory are being used in install_copy() and
install_alternative() then there are no dependencies to these files.

This is being solved by storing all dependencies of the targetinstall
stage in a seperate state/<package>.deps file.
These state/<package>.deps files are included in the top-level ptxdist makefile
and this makes ptxdist properly handle the dependencies.

Signed-off-by: Remy Bohmer <linux@bohmer.net>
Signed-off-by: Bart vdr. Meulen <bartvdrmeulen@gmail.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>