[libptxdist] ptxd_replace_magic: don't replace unset MAGICs
This patch changes the semantics of the function.

Originally all @MAGIC@ are replaced with their environemnt value. Even
if MAGIC is unset in the env, resulting in @MAGIC@ to be removed from
the output.

With this patch ptxd_replace_magic will only replace @MAGIC@ if MAGIC is
set in the environment.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>