[ptxdist] add new packet x-load
x-load is a first level bootloader for Texas Intruments OMAP
microprocessors (similar to the AT91Bootstrap application).

This bootstrap package is required on TI OMAP boards where
dataflash/nandflash or USB/RS232 is used to boot from. In
any of these cases the internal bootROM of the OMAP core is
used to boot the board with.

Signed-off-by: Stephan Linz <linz@li-pro.net>
[mkl: removed trailing newline]
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>