[ptxdist] stricter search algorithm for install_lib
Until now the $(install_lib) mechanism is searching for lib<bla>.so*
anywhere in /lib,/usr/lib. In libpv we have a libpv.so*, but also a
python module deeper in the tree that is also named libpv.so, so the
find triggers twice and the check for a file fails.

On the other hand in mysql the libs are in a subdirectory. To make
both work we now match the whole path to {/lib,/usr/lib}/${libname}.so*
and add the subdirectories to libname.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>