[ptxd_make_00-init] prevent endless loop for collections with base platform
When a base platform is used we try to create a list of packages built
by that platform. This is done by faking that platform and calling
make.
With the changes made in 7dad70a7a44da6867c5b370a7298b8ac11241547
PTXDIST_BASE_PLATFORMDIR is already defined at this point. The result
is an endless recursion.
Unset PTXDIST_BASE_PLATFORMDIR before calling make and improve the
faking.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>