fix failing cd in prepare stage if <PKG>_CONF_TOOL=NO and empty extract
In case the extract stage is empty and <PKG>_CONF_TOOL is set to NO, then
the prepare stage fails. First a cd was done to the build_dir, but that
does not exist when the extract is empty. This makes the prepare stage fail
while we told it with <PKG>_CONF_TOOL=NO to skip that stage

Signed-off-by: Remy Bohmer <linux@bohmer.net>
[mol: fix indent]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>