glib: version bump 2.26.1 -> 2.28.0
There is a new stable series out there, so this patch bumps to 2.28. As
this is almost identical with the former 2.27.93 beta version, remove
the experimental version for now. The result was build-tested with the
AllYes BSP.

We leave the experimental mechanism in the rules file: there is
currently no experimental series released yet, but this will happen
again in the future.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>