mplayer: cleanup advanced options
Neither the currently by ptxdist built 1.0.0rc2 nor the newer
1.0.0rc4 have a --enable-pxa switch, so remove this completely.
For testing on iwmmxt we now have a explicit switch and do not
need to depend on pxa anymore. This leaves a menu with only
one entry left, so remove the menu and put the iwmmxt extension
switch along with the other options.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>