lndir: add support for lndir:// URL
Use lndir for local external projects referenced by lndir:// URL

lndir will create a shadow directory tree of the source tree populated with
symlinks of all the files contained in the source tree. Building from the
lndir created shadow tree will result in all files created by the build to
be created in the shadow tree and leaving the source tree prestine. This is
useful when building for multiple archs.

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>