git-ptx-patches: Allow patches before base tag
The patch series I have for the Linux kernel includes incremental patches
from upstream like

patch-2.6.33.1.bz2
patch-2.6.33.1-2.bz2

followed by a patch series for my board.
But ptxdist messes this up when applying the patch series with git.

This patch allows one to mark the base tag in the series with:
#tag:base --start-number 1

Other arbitrary tags can also be put in the series:
#tag:for-upstream --start-number 5

When using 'git ptx-patches' to regenerate the series, only the patches
between "#tag:base" and the next "#tag:*" or EOF in the series are
regenerated. All others in the series are left untouched. To use a
different tag than the default tag base, use -t tagname.

e.g.:
# generated by git-ptx-patches
patch-2.6.33.1.bz2
patch-2.6.33.1-2.bz2
patch-2.6.33.2-3.bz2
#tag:for-upstream --start-number 1
0001-for-upstream-1.patch
0002-for-upstream-2.patch
0003-for-upstream-3.patch
#tag:base --start-number 4
0004-foo.patch
0005-bar.batch
# 25a3096f40bc8d841c8bf3ec24dd101f  - git-ptx-patches magic

Any options included on a #tag:* line are passed to git format-patch when
generating the patches for that section.

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>