boost: fix crosscompiling
boost.threads would generate a host lib rather than a target lib.

Following instructions for crosscompiling here:
http://www.boost.org/boost-build2/doc/html/bbv2/tasks/crosscompile.html

It appears that as long as the version string in user-config.jam matches
the toolset version string, it works.

for example, a user-config.jam with:
using gcc : foobar : arm-none-linux-gnueabi-g++ ;

would be invoked with this to crosscompile with version foobar:
bjam toolset=gcc-foobar

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>