python: use host-python when running regen script
The regen script is called with the python toplevel directory in the
path, so the cross compiled python is called from the script. Fix that
by removing the current dir from PATH.

I'm not sure why this issue hasn't been observed yet, probably because
the regen script was never run. The question remains why it *is* run
now...

Reworked the patch descriptions while being there.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>