Add the userspace RCU library
From its README:
This data synchronization library provides read-side access which scales
linearly with the number of cores. It does so by allowing multiples copies
of a given data structure to live at the same time, and by monitoring the data
structure accesses to detect grace periods after which memory reclamation is
possible.

Supported architectures are: x86, PPC and ARM.

Signed-off-by: Juergen Beisert <jbe@pengutronix.de>