host-e2fsprogs: Fix build on Darwin
The O_DIRECT flag in e2fsprogs is relatively new. MacPorts is still using the
old version of e2fsprogs. However, this patch from the upstream bug tracker
fixes the build of e2fsprogs.

Strictly direct I/O is not required for ptxdist because we're working on
images only. So removing O_DIRECT here would also work, but this solution is
cleaner.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>