opkg-utils: use toolchain provided ar
On OS X (and maybe some other systems) the default provided ar is not a
GNU version and does not understand all the required switches.

To circumvent this fact always use the ar from cross tools which is
proven to be a GNU binutils version (at least with OSELAS.Toolchain).

Signed-off-by: Andreas Bießmann <andreas@biessmann.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>