Setup vim folding in the logfile automatically
This patch adds the modeline

        # vim: set fdm=marker fdl=1:

to the Logfile. There's already a modeline for Emacs in the file, so it
shouldn't be too annoying for users that don't use vim.

The first setting tells vim to recognize the fold markers '{{{' and
'}}}' in the logfile (which doesn't have any file type in vim).

However, with this setting vim automatically closes all folds, including
the first '{{{' which only is for Emacs. That's why I add fdl=1, which
results in vim opening the file visible one fold per ptxdist run.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>