Add host-elf-h-compat package
This package should provide <elf.h> on host operating systems that have
no /usr/include/elf.h. The header is copied from glibc, modified a bit
to remove GNU dependencies and is sufficient to compile the Linux kernel
and module-init-tools.

The package itself only checks if elf.h is present. If yes, it does
nothing on "make install". If not, it installs its own copy of elf.h in
"make install" to the include directory set in autoconf.

We add this package as base package so that it gets built before the
packages which need it. Also, the package only gets required when
ptxdist is executed not on Linux (per uname -s output). This makes
maintaining the dependencies easier than requiring it in any packages
that need elf.h.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>