platforms: Make it possible to create a uRamdisk with an initramfs
This patch adds a choice to make it possible to create the uRamdisk file
not as initrd but as initramfs. One advantage is that you don't have to
know the maximum size in advance.

The patch also beautifies the output a bit (all other images are created
silently, so we can also run this mkimage silently).

Signed-off-by: Bernhard Walle <walle@corscience.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>