configure.ac: Check for GNU tar
ptxd_make_world_extract uses the option --remove-files which is a GNU
extension to tar.

Because GNU tar is not in GNU coreutils, I changed the GNU_COREUTILS
macro to GNU_TOOL and added a second parameter: the package where the
tool is found. I think that's better than providing a new macro just for
GNU tar.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>