ptxdist: Check for GNU readlink
ptxdist heavily uses "readlink -f" which only works with GNU readlink.
So add the usual check here.

However, because "readlink -f" is also used before the symlinks in the
ptxdist directory is added to the PATH, that "readlink -f" call is
replaced by a bash-only implementation. This one is inspired from
http://mailman.archlinux.org/pipermail/pacman-dev/2009-February/008130.html.

Tested on Mac OS 10.7.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
[mol: fixed some more quoting]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>