Add flashbench tool
This patch adds the flashbench tool from Linaro Project
git://git.linaro.org/people/arnd/flashbench.git. See also
http://lwn.net/Articles/428584/. Because there is no tarball, I
exported the current git repo and put it on my server.

The default maximum buffer size of 64 MiB (the buffer is allocated 4
times) was too large for my device, so I patched the tool and added an
option in the ptxdist menu.

Tested with a custom OMAP3 board which is similar to the Devkit8000.

Signed-off-by: Bernhard Walle <walle@corscience.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>