add /etc/timezone and replace content according to /etc/localtime
Additionally to /etc/localtime which points to one zoneinfo file in
/usr/share/zoneinfo there is /etc/timezone containing the system wide
zone. Debian updates both in parallel and systemd uses this too. It's
an easy way to find out about the current system time zone. Rules are
extended in a way you only have to put e.g. 'Europe/Berlin' to ptxdist
menuconfig and symlink /etc/localtime and file /etc/timezone are
created at build time.

This patch is a suggestion which works for me, open for discussion.

Signed-off-by: Alexander Dahl <post@lespocky.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>