configure.ac: Check for GNU find
One of the specialities of GNU find over BSD find is that you can omit
the path before providing options. This means that "find -name bla" is
equivalent with "find . -name bla". Busybox exposes that behaviour.

So check for GNU find to be on the safe side, and also add the symlink
in the installation directory once we found it.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>