fontconfig: Fix Darwin build
Upstream solved it completely different
http://cgit.freedesktop.org/fontconfig/commit/?id=2a6b235ff6d2750171e8dff7cfdfe3bccb0f630e
now, so after updating fontconfig the patch will be obsolete. Applying
that patch is not possible, however, since I would need to re-create the
tarball.

However, I don't want to update fontconfig since it touches the whole
Xorg stack and I have no possibility to test that.

The first upstream version that contains the release is 2.9.0.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>