ptxd_make_xpkg_pkg: add possibility to specify user/group as names
Before this patch, the user/group information has to specified as numeric
values. With this patch it's possible to write their names, too. E.g.:

-       @$(call install_copy, apache2, 12, 102, 0755, $(PTXCONF_APACHE2_SERVERROOT))
+       @$(call install_copy, apache2, www, www, 0755, $(PTXCONF_APACHE2_SERVERROOT))

The files /etc/passwd and /etc/group are used to lookup the numeric values.
As always the files in the BSP's projectroot have precedence over the generic
files in ptxdist.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>