readd _SERIES option for u-boot and add it for at91bootstrap (partially reverts dac31c4)
In dac31c4 U_BOOT_SERIES was removed with the explanation it would not
be used in ptxdist anymore. This is not true. In
./scripts/lib/ptxd_make_world_patchin.sh and
./rules/post/ptxd_make_world_patchin.make there's a mechanism which
automatically uses an option PACKAGENAME_SERIES if present. This can
be used to have different series file in one patches directory for a
given package. Removing U_BOOT_SERIES broke our BSP using this
mechanism.

Signed-off-by: Alexander Dahl <post@lespocky.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>