libselinux: fix parallel building issue
Analogue to: "245f79f libsemanage: don't add 'all' to the make options"

It's implicit if nothing else is given, and 'pywrap' depends on anything that
is needed.  If both 'all' and 'pywrap' are explicitly added libsemanage.so.1
and it files a build twice. This can cause random build error for parallel
building.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>