ptxd_install_shared: Fix missing links when ${src} is a symlink itself
When "${src}" is a symlink to some other directory, then not all
symbolic links to the library are installed.

In my case I'm using a ct-ng-based toolchain where
${TOOLCHAIN_BASE}/arm-unknown-linux-gnueabi/lib/ is a symbolic link to
sysroot/lib/. As a result, install_copy_toolchain_lib with libstdc++.so
doesn't install the libstdc++.so.6 symbolic link and so C++ programs
don't start.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>