Fix (host) gcc wrapper on Gentoo Linux
On Gentoo, the host compiler can only be executed if basename(argv[0])
is 'gcc' (or 'g++' etc.), not if it's 'gcc.real'. Then it invokes
gcc-config and fails:

 | % ln -s /usr/bin/gcc gcc.real
 | % ./gcc.real --version
 |
 |  * gcc-config: Could not get portage CHOST!
 |  * gcc-config: You should verify that CHOST is set in one of these places:
 |  * gcc-config:  - //etc/portage/make.conf
 |  * gcc-config:  - active environment
 | gcc-config: error: could not get compiler binary path: No such file or directory

I'm not a Gentoo expert (I'm just forced to use it for that purpose ;-)),
so I cannot explain further details.

This patch makes ptxdist working on Gentoo by changing the symbolic link
from <name>.real to real/<name>, i.e. moving the final toolchain symlink
to some other directory, keeping the basename.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
[mol: use shell magic instead of dirname/basename]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>