libpcap: version bump to 1.4.0
Bump libpcap version to 1.4.0.

Changes:
  - Add --enable-bluetooth configure parameter only if LIBPCAP_BLUETOOTH
    is set.
  - Add --disable-universal to disable build for Mac OS.
  - Add --enable-shared to build shared library.

Both are default parameter but we avoid a change of default behaviour
here.

Remove patch 0001-add-CFLAGS-CPPFLAGS-and-LDFLAGS-to-Makefile.in.patch:
  - This patch is already mainline see:
    https://github.com/the-tcpdump-group/libpcap/commit/3a089e628c2ef39efdb64c53a7566052917ae974

Update other patches to current version and put a signed-off by me in
the commit message.

Change 0002-pcap-config-add-SYSROOT-support.patch patch to add SYSROOT
dir only in includedir and libdir variable, so it's easier to maintain.

Tested with wireshark 1.4.0 and tcpdump 4.1.1

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>