ptxd_make_get: support svn urls
In some circumstances it is important to fetch source code
directly from a repository, for example for software development
and/or evaluation. There are some few software projects that
still stay on Subversion, especially vendor specific projects.

In very rare cases the project maintainer will never provide
release tarballs and you have to fetch the Subversion release
branch (for example the eglibc project).

Signed-off-by: Stephan Linz <linz@li-pro.net>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>