Fix problem with finding platforms/*.in
ptxdist 2013.08 doesn't find my platforms/u-boot.in file. This is
because it looks for ":/path/to/bsp/platforms" instead of the variant
without colon.

Following shell script demonstrates that:

    #!/bin/bash

    oldifs="$IFS"
    IFS=: a="1:2:3:" b="4:5:6:"
    c=(${a}:${b})
    d=(${a} ${b})
    IFS="$oldifs"
    echo ${c[@]} # 1 2 3 :4 5 6
    echo ${d[@]} # 1 2 3 4 5 6

So only the variant without a ':' is correct.

The problem was introduced with 51c5786c16a9a3508af9b6cc69947a2f3b8bc7cf
on 20th June 2013. So I think it was also in 2013.07 although I didn't
test.(

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>