net-snmp: check mib config by default
If you compile additional MIB modules --enable-mib-config-checking
checks for conlicts in the options you provided. Without setting this, a
warning would be given, but the build would proceed with deactivated
options and silently not building a desired module would be the result.
Setting this the build fails and the user can correct the chosen MIB
modules.

Signed-off-by: Alexander Dahl <post@lespocky.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>