udev: standalone variant fails to build with newer toolchains
udev cannot be linked due to the use of the 'mq_getattr' call which requires
librt to link successfully.

The configure script already checks for 'clock_gettime' in order to add librt
to the list of required libraries. This is valid only (according to the manual
page) for glibc version before 2.17. With the new glibc-2.18 in the 2013.12
toolchain librt isn't required anymore for 'clock_gettime' and that is why
librt is omitted and linking fails.

This change uses the upstream patch from the systemd repository.

Signed-off-by: Juergen Beisert <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>