php: version bump 5.3.10 -> 5.4.24
php: version bump 5.3.10 -> 5.4.24

Fix compiling of PHP with sqlite3 support on ptxdist 2012.04.0.

Fixes a lot of security flaws and bugs.

Changelog:
http://www.php.net/ChangeLog-5.php#5.4.24

PHP: Removal of curl-wrappers.
This feature has been marked as experimental for years and never turned to stable.
Most people using it just do so by accident.
This experimental feature should be moved in a new PECL extension for improvements until it gets stable.
https://wiki.php.net/rfc/curl-wrappers-removal-rfc

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>